refactor: :hammer: introduce AbstractUsersTestCase for shared setup in user-related integration tests
