+++ tests/Integration/Application/Users/Actions/DeleteUserActionTest.php
@@ -4,47 +4,13 @@ declare(strict_types=1);
 
 namespace Tests\Integration\Application\Users\Actions;
 
-use App\Domain\User\UserRepository;
-use Psr\Container\ContainerExceptionInterface;
-use Psr\Container\NotFoundExceptionInterface;
-use Tests\Integration\AbstractIntegrationTestCase;
-use Tests\Support\Factories\UserFactory;
+use Tests\Integration\Application\Users\AbstractUsersTestCase;
 
 /**
  * Integration tests for DeleteUserAction.
  */
-final class DeleteUserActionTest extends AbstractIntegrationTestCase
+final class DeleteUserActionTest extends AbstractUsersTestCase
 {
-    /**
-     * The user repository instance.
-     */
-    private UserRepository $users;
-
-    /**
-     * The user factory instance.
-     */
-    private UserFactory $userFactory;
-
-    /**
-     * Sets up the test dependencies before each test.
-     *
-     * @return void
-     * @throws ContainerExceptionInterface
-     * @throws NotFoundExceptionInterface
-     */
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $this->users = $this->app
-            ->getContainer()
-            ->get(UserRepository::class);
-
-        $this->userFactory = new UserFactory(
-            $this->users,
-            $this->faker
-        );
-    }
 
     /**
      * Asserts that a 204 response with an empty body is returned when the user exists.
@@ -66,7 +32,7 @@ final class DeleteUserActionTest extends AbstractIntegrationTestCase
     {
         $user = $this->userFactory->create();
 
-        $this->users->delete($user->getId());
+        $this->userRepository->delete($user->getId());
 
         $response = $this->request('DELETE', '/api/v1/users/' . $user->getId());
