+++ tests/Integration/Application/Users/Actions/UpdateUserActionTest.php
@@ -4,47 +4,13 @@ declare(strict_types=1);
 
 namespace Tests\Integration\Application\Users\Actions;
 
-use App\Domain\User\UserRepository;
-use Psr\Container\ContainerExceptionInterface;
-use Psr\Container\NotFoundExceptionInterface;
-use Tests\Integration\AbstractIntegrationTestCase;
-use Tests\Support\Factories\UserFactory;
+use Tests\Integration\Application\Users\AbstractUsersTestCase;
 
 /**
  * Integration tests for UpdateUserAction.
  */
-final class UpdateUserActionTest extends AbstractIntegrationTestCase
+final class UpdateUserActionTest extends AbstractUsersTestCase
 {
-    /**
-     * The user repository instance.
-     */
-    private UserRepository $users;
-
-    /**
-     * The user factory instance.
-     */
-    private UserFactory $userFactory;
-
-    /**
-     * Sets up the test dependencies before each test.
-     *
-     * @return void
-     * @throws ContainerExceptionInterface
-     * @throws NotFoundExceptionInterface
-     */
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $this->users = $this->app
-            ->getContainer()
-            ->get(UserRepository::class);
-
-        $this->userFactory = new UserFactory(
-            $this->users,
-            $this->faker
-        );
-    }
 
     /**
      * Asserts that a 200 response containing the updated user data is returned when the user exists.
@@ -83,7 +49,7 @@ final class UpdateUserActionTest extends AbstractIntegrationTestCase
     {
         $user = $this->userFactory->create();
 
-        $this->users->delete($user->getId());
+        $this->userRepository->delete($user->getId());
 
         $response = $this->request('PUT', '/api/v1/users/' . $user->getId(), [
             'first_name' => 'Ghost',
